+++ src/typescript.ts
@@ -101,6 +101,7 @@ export class TypeScriptProject extends NodeProject {
   public readonly docsDirectory: string;
   public readonly eslint?: Eslint;
   public readonly jest?: Jest;
+  public readonly tsconfig?: TypescriptConfig;
 
   /**
    * The directory in which the .ts sources reside.
@@ -141,10 +142,11 @@ export class TypeScriptProject extends NodeProject {
 
     // by default, we first run tests (jest compiles the typescript in the background) and only then we compile.
     const compileBeforeTest = options.compileBeforeTest ?? false;
+
     if (compileBeforeTest) {
-      this.addBuildCommand('yarn compile', 'yarn test');
+      this.addBuildCommand(`${this.runScriptCommand} compile`, `${this.runScriptCommand} test`);
     } else {
-      this.addBuildCommand('yarn test', 'yarn compile');
+      this.addBuildCommand(`${this.runScriptCommand} test`, `${this.runScriptCommand} compile`);
     }
     this.start?.addEntry('build', {
       desc: 'Full release build (test+compile)',
@@ -155,11 +157,11 @@ export class TypeScriptProject extends NodeProject {
       this.addScript('package',
         'rm -fr dist',
         'mkdir -p dist/js',
-        'yarn pack',
+        `${this.packageManager} pack`,
         'mv *.tgz dist/js/',
       );
 
-      this.addBuildCommand('yarn run package');
+      this.addBuildCommand(`${this.runScriptCommand} package`);
 
       this.start?.addEntry('package', {
         desc: 'Create an npm tarball',
@@ -195,7 +197,7 @@ export class TypeScriptProject extends NodeProject {
     };
 
     if (!options.disableTsconfig) {
-      new TypescriptConfig(this, {
+      this.tsconfig = new TypescriptConfig(this, {
         include: [`${this.srcdir}/**/*.ts`],
         exclude: [
           'node_modules',